type adjustments

+++ test/test_radial_branches.jl
@@ -64,7 +64,7 @@ end
         a_mat = IncidenceMatrix(sys)
         Y = Ybus(sys; network_reductions = [NetworkReductionTypes.RADIAL])
         rb = get_network_reduction(Y)
-        leaf_buses = Int64[]
+        leaf_buses = Int[]
         for i in keys(rb.bus_reduction_map)
             append!(leaf_buses, collect(rb.bus_reduction_map[i]))
         end
@@ -82,5 +82,5 @@ end
     rr = get_radial_reduction(a_mat; exempt_buses = [16, 17])
     @test !haskey(rr.reverse_bus_search_map, 16)
     @test !haskey(rr.reverse_bus_search_map, 17)
-    #@test isa(PTDF(sys; network_reduction = rr), PTDF)     #TODO - look at this test once PTDF works; I think I have captured the functionality with other added tests 
+    #@test isa(PTDF(sys; network_reduction = rr), PTDF)     #TODO - look at this test once PTDF works; I think I have captured the functionality with other added tests
 end
